Partially addressed code review, moved SMOTENC to _smote module.
