+++ imblearn/over_sampling/__init__.py
@@ -8,7 +8,7 @@ from ._random_over_sampler import RandomOverSampler
 from ._smote import SMOTE
 from ._smote import BorderlineSMOTE
 from ._smote import SVMSMOTE
-from ._smote_nc import SMOTENC
+from ._smote import SMOTENC
 
 __all__ = ['ADASYN', 'RandomOverSampler',
            'SMOTE', 'BorderlineSMOTE', 'SVMSMOTE', 'SMOTENC']
